Created a Person superclass
