+++ src/main/java/seedu/duke/Duke.java
@@ -1,7 +1,5 @@
 package seedu.duke;
 
-import java.util.Scanner;
-
 public class Duke {
 
     /**
