+++ src/main/java/seedu/duke/assets/Patient.java
@@ -1,47 +1,12 @@
 package seedu.duke.assets;
 
-public class Patient {
-    String nric;
-    String fullName;
-    int age;
-    char gender;
-    String address;
-    String dob;
+public class Patient extends Person {
     String dateAdmission;
 
     public Patient(String nric, String fullName, int age, char gender, String address, String dob,
                    String dateAdmission) {
-        this.nric = nric;
-        this.fullName = fullName;
-        this.age = age;
-        this.gender = gender;
-        this.address = address;
-        this.dob = dob;
+        super(nric, fullName, age, gender, address, dob);
         this.dateAdmission = dateAdmission;
-    }
-
-    public String getNric() {
-        return nric;
-    }
-
-    public String getFullName() {
-        return fullName;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public char getGender() {
-        return gender;
-    }
-
-    public String getAddress() {
-        return address;
-    }
-
-    public String getDob() {
-        return dob;
     }
 
     public String getDateAdmission() {
